refact(test): merge support.js and helper.js
